pyproject.toml
[project]
name = "wplt"
version = "1.0.0"
authors = [
    { name="Emil Wagman", email="emilwagman@orpose.com" }
]
description = "[To be written]"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy >= 1.25.0",
    "Pillow >= 10.1.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[testpypi]
username = "__token__"
password = "pypi-AgENdGVzdC5weXBpLm9yZwIkOGM5MzMwMjAtN2I0OC00MWQ3LWE5YTktZjg4Zjk2MGM3ZGZmAAIqWzMsIjg5NGVmZTRlLTIzODUtNDZkNy04OWZiLTM1ZWE1NjljYzBhZSJdAAAGIClzBA61vCZpRBP9RmabZNmGafFlNF60K4x6E_Is7u4K"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"